More fiddling, split 'LumigatorClient' out from the ApiClient (in files)
