+++ lumigator/python/mzai/sdk/tests/test_completions.py
@@ -1,7 +1,7 @@
 import json
 
 
-def test_get_vendors_ok(mock_requests_response, mock_requests, lumi_client):
+def test_get_vendors_ok(mock_requests_response, lumi_client):
     mock_requests_response.status_code = 200
     mock_requests_response.json = lambda: json.loads('["openai", "mistral"]')
 
@@ -13,7 +13,7 @@ def test_get_vendors_ok(mock_requests_response, mock_requests, lumi_client):
     assert vendors[1] == "mistral"
 
 
-def test_get_vendors_none(mock_requests_response, mock_requests, lumi_client):
+def test_get_vendors_none(mock_requests_response, lumi_client):
     mock_requests_response.status_code = 200
     mock_requests_response.json = lambda: json.loads("[]")
